requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
prophet>=1.1.5
scikit-learn>=1.3.0
python-dateutil>=2.8.0
requests>=2.31.0
jpholiday>=0.1.8
